+++ src/components/modal/input_description/handlers.ts
@@ -1,0 +1,12 @@
+import restaurantStateStore from "../../../store/RestaurantStateStore";
+import removeHTMLElementByClassName from "../../../utils/removeHTMLElementByClassName";
+
+export const descriptionEventHandler = (description: HTMLElement) => {
+  description.addEventListener("input", (event) => {
+    if (event.target instanceof HTMLTextAreaElement) {
+      const inputDescription = event.target.value;
+      removeHTMLElementByClassName("invalid_description");
+      restaurantStateStore.setDescription(inputDescription);
+    }
+  });
+};
