+++ src/components/modal/select_category/handlers.ts
@@ -1,0 +1,23 @@
+import restaurantStateStore from "../../../store/RestaurantStateStore";
+import { Icategory } from "../../../types";
+import removeHTMLElementByClassName from "../../../utils/removeHTMLElementByClassName";
+
+export const categoryChange = () => {
+  document.addEventListener("DOMContentLoaded", () => {
+    const select = document.getElementById("category");
+
+    if (select) {
+      selectEventHandler(select);
+    }
+  });
+};
+
+export const selectEventHandler = (select: HTMLElement) => {
+  select.addEventListener("change", (event) => {
+    if (event.target instanceof HTMLSelectElement) {
+      const selectedValue = event.target.value;
+      removeHTMLElementByClassName("invalid_category");
+      restaurantStateStore.setCategory(selectedValue as Icategory);
+    }
+  });
+};
