+++ src/components/modal/select_distance/handlers.ts
@@ -1,0 +1,24 @@
+import restaurantStateStore from "../../../store/RestaurantStateStore";
+import { Idistance } from "../../../types";
+import removeHTMLElementByClassName from "../../../utils/removeHTMLElementByClassName";
+
+export const distanceChange = () => {
+  document.addEventListener("DOMContentLoaded", () => {
+    const select = document.getElementById("distance");
+
+    if (select) {
+      selectDistanceEventHandler(select);
+    }
+  });
+};
+
+export const selectDistanceEventHandler = (select: HTMLElement) => {
+  select.addEventListener("change", (event) => {
+    if (event.target instanceof HTMLSelectElement) {
+      const selectedValue = Number(event.target.value) as Idistance;
+      removeHTMLElementByClassName("invalid_distance");
+      restaurantStateStore.setDistance(selectedValue);
+      return selectedValue;
+    }
+  });
+};
